pulled the two subclasses into their own files
